+++ frontend/src/pages/my/MyPage.tsx
@@ -12,8 +12,8 @@ import Comment from '../../assets/icons/comment.svg?react'
 import { formatKoreanNumber } from '../../utils/format'
 import { useParams } from 'react-router-dom'
 import { useGetChannel } from '../../hooks/my/useGetChannel'
-import { mapResponseToProfile } from '../../mappers/profile.ts/mapResponseToProfile'
-import { mapResponseStatCard } from '../../mappers/profile.ts/mapResponsetoStatCard'
+import { mapResponseToProfile } from '../../lib/mappers/profile/mapResponseToProfile'
+import { mapResponseStatCard } from '../../lib/mappers/profile/mapResponsetoStatCard'
 import { useEffect } from 'react'
 import { setMockLogin } from '../../utils/loginMock'
 import { Skeleton } from './_components/Skeleton'
